refac: rename project to kstash
