+++ README.md
@@ -1,13 +1,13 @@
-# fdstash
+# kstash
 
 A Python library for transparently linking large message attributes using remote object storage.
 
-[![CI](https://github.com/ccortezia/fdstash/actions/workflows/main.yml/badge.svg)](https://github.com/ccortezia/fdstash/actions/workflows/main.yml)
+[![CI](https://github.com/ccortezia/kstash/actions/workflows/main.yml/badge.svg)](https://github.com/ccortezia/kstash/actions/workflows/main.yml)
 
 ## Installation
 
 ```bash
-pip install fdstash
+pip install kstash
 ```
 
 ## Requirements
@@ -27,16 +27,16 @@ conn.create_bucket(Bucket="stashes")
 ```
 
 ```python
-import fdstash
+import kstash
 
 # Process A: Sends a message containing the address of a stashed payload.
 context = {"as_of": "today", "bin": b"0" * 1024 * 512}
-stash = fdstash.create("context", context, namespace="stashes")
+stash = kstash.create("context", context, namespace="stashes")
 message = {"context": str(stash.address), "command": "gen_report"}  
 assert str(stash.address) == "s3://stashes/context.c6ab205fe81dcad3eec3ab48b96b0618"
 
 # Process B: Rebuilds the message from the stash's address.
-loaded_stash = fdstash.retrieve(message["context"])
+loaded_stash = kstash.retrieve(message["context"])
 assert loaded_stash == stash
 ```
 
@@ -44,7 +44,7 @@ assert loaded_stash == stash
 
 Use `stash.share()` to produce a short-lived HTTPS link to the stash. 
 
-See `fdstash.Config.share_ttl_sec` for configuration details.
+See `kstash.Config.share_ttl_sec` for configuration details.
 
 ```python
 import time 
@@ -54,26 +54,26 @@ shared_link = stash.share(ttl_sec=5)
 
 # Process B: Rebuilds the message from the shared link.
 time.sleep(3)
-loaded_stash = fdstash.retrieve(shared_link)
+loaded_stash = kstash.retrieve(shared_link)
 assert loaded_stash == shared_stash
 
 # Process B: Fails to retrieve the stash from an expired link.
 time.sleep(3)
-loaded_stash = fdstash.retrieve(shared_link)
+loaded_stash = kstash.retrieve(shared_link)
 # Error: StashNotFound, share link expired
 ```
 
 ## Inline Data Optimization
 
-`fdstash.create()` embeds small data in the stash's address.
+`kstash.create()` embeds small data in the stash's address.
 
-See `fdstash.Config.max_inline_len` for configuration details.
+See `kstash.Config.max_inline_len` for configuration details.
 
 ```python
-import fdstash
-stash = fdstash.create("colorcode", 12, namespace="stashes")
+import kstash
+stash = kstash.create("colorcode", 12, namespace="stashes")
 assert str(stash.address) == "inline://stashes/colorcode?data=DA%3D%3D"
-loaded_stash = fdstash.retrieve(stash.address)
+loaded_stash = kstash.retrieve(stash.address)
 assert loaded_stash == stash
 loaded_stash.data
 ```
@@ -82,14 +82,14 @@ loaded_stash.data
 
 Backends can be disabled to address specific deployment or test scenarios.
 
-See `fdstash.Config.backends` for more details.
+See `kstash.Config.backends` for more details.
 
 ```python
-import fdstash
-config = fdstash.Config(backends=["mem"])
-stash = fdstash.create("object", {"data": 123}, config=config)
+import kstash
+config = kstash.Config(backends=["mem"])
+stash = kstash.create("object", {"data": 123}, config=config)
 assert str(stash.address) == "mem://default/object.34472d91b2f84052bf26d4eaa862ef86"
-loaded_stash = fdstash.retrieve(stash.address, config=config)
+loaded_stash = kstash.retrieve(stash.address, config=config)
 assert loaded_stash == stash
 ```
