+++ src/kstash/backend_base.py
