+++ src/kstash/backend_inline.py
