+++ src/kstash/backend_mem.py
