+++ src/kstash/exceptions.py
